Split code from lib.php for returning grades into a separate file for calling from multiple locations.
Working to enable the ability to be able to post to a file within the LTIProvider module with a SourcedId and a tool URL to get grades on demand.

added 'currentcourse' to navigation nodes to be deleted to accomodate Moodle 2.4 changes.
